requirement.txt
loguru==0.5.3
mysql-connector==2.2.9
numpy==1.19.4
pandas==1.1.4
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
